organise machines
